requirements.txt

beautifulsoup4==4.9.3
bs4==0.0.1

psycopg2-binary==2.9.3

requests==2.25.1
